[DependencyInjection] Add Kernel and Bundle infrastructure for HTTP-less DI-powered apps
