+++ src/Symfony/Bundle/FrameworkBundle/Console/Application.php
@@ -194,7 +194,7 @@ class Application extends BaseApplication implements ContainerAwareInterface
 
         foreach ($this->kernel->getBundles() as $bundle) {
             if ($bundle instanceof Bundle
-                && \method_exists($bundle, 'registerCommands')
+                && method_exists($bundle, 'registerCommands')
                 && Bundle::class !== new \ReflectionMethod($bundle, 'registerCommands')->getDeclaringClass()->getName()
             ) {
                 trigger_deprecation('symfony/framework-bundle', '8.1', 'Overriding the "%s::registerCommands()" method in "%s" is deprecated, use the "#[AsCommand]" attribute or the "console.command" service tag instead.', Bundle::class, get_debug_type($bundle));
