+++ src/Symfony/Bundle/FrameworkBundle/composer.json
@@ -21,14 +21,14 @@
         "ext-xml": "*",
         "symfony/cache": "^7.4|^8.0",
         "symfony/config": "^7.4.4|^8.0.4",
-        "symfony/dependency-injection": "^7.4.4|^8.0.4",
+        "symfony/dependency-injection": "^8.1",
         "symfony/deprecation-contracts": "^2.5|^3",
         "symfony/error-handler": "^7.4|^8.0",
         "symfony/event-dispatcher": "^7.4|^8.0",
         "symfony/filesystem": "^7.4|^8.0",
         "symfony/finder": "^7.4|^8.0",
         "symfony/http-foundation": "^7.4|^8.0",
-        "symfony/http-kernel": "^7.4|^8.0",
+        "symfony/http-kernel": "^8.1",
         "symfony/polyfill-mbstring": "^1.0",
         "symfony/polyfill-php85": "^1.33",
         "symfony/routing": "^7.4|^8.0",
