+++ src/Symfony/Component/HttpKernel/Bundle/AbstractBundle.php
@@ -11,51 +11,23 @@
 
 namespace Symfony\Component\HttpKernel\Bundle;
 
-use Symfony\Component\Config\Definition\Configurator\DefinitionConfigurator;
-use Symfony\Component\DependencyInjection\Container;
-use Symfony\Component\DependencyInjection\ContainerBuilder;
-use Symfony\Component\DependencyInjection\Extension\ConfigurableExtensionInterface;
 use Symfony\Component\DependencyInjection\Extension\ExtensionInterface;
-use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
+use Symfony\Component\DependencyInjection\Kernel\AbstractBundle as BaseAbstractBundle;
 
 /**
  * A Bundle that provides configuration hooks.
  *
  * @author Yonel Ceruto <yonelceruto@gmail.com>
  */
-abstract class AbstractBundle extends Bundle implements ConfigurableExtensionInterface
+abstract class AbstractBundle extends Bundle
 {
-    protected string $extensionAlias = '';
-
-    public function configure(DefinitionConfigurator $definition): void
-    {
-    }
-
-    public function prependExtension(ContainerConfigurator $container, ContainerBuilder $builder): void
-    {
-    }
-
-    public function loadExtension(array $config, ContainerConfigurator $container, ContainerBuilder $builder): void
-    {
-    }
-
     public function getContainerExtension(): ?ExtensionInterface
     {
-        if ('' === $this->extensionAlias) {
-            $this->extensionAlias = Container::underscore(preg_replace('/Bundle$/', '', $this->getName()));
-        }
-
-        return $this->extension ??= new BundleExtension($this, $this->extensionAlias);
+        return BaseAbstractBundle::getContainerExtension();
     }
 
     public function getPath(): string
     {
-        if (!isset($this->path)) {
-            $reflected = new \ReflectionObject($this);
-            // assume the modern directory structure by default
-            $this->path = \dirname($reflected->getFileName(), 2);
-        }
-
-        return $this->path;
+        return BaseAbstractBundle::getPath();
     }
 }
