+++ src/Symfony/Component/HttpKernel/Bundle/Bundle.php
@@ -13,40 +13,16 @@ namespace Symfony\Component\HttpKernel\Bundle;
 
 use Symfony\Component\Console\Application;
 use Symfony\Component\DependencyInjection\Container;
-use Symfony\Component\DependencyInjection\ContainerBuilder;
-use Symfony\Component\DependencyInjection\ContainerInterface;
 use Symfony\Component\DependencyInjection\Extension\ExtensionInterface;
+use Symfony\Component\DependencyInjection\Kernel\AbstractBundle as BaseAbstractBundle;
 
 /**
  * An implementation of BundleInterface that adds a few conventions for DependencyInjection extensions.
  *
  * @author Fabien Potencier <fabien@symfony.com>
  */
-abstract class Bundle implements BundleInterface
+abstract class Bundle extends BaseAbstractBundle implements BundleInterface
 {
-    protected string $name;
-    protected ExtensionInterface|false|null $extension = null;
-    protected string $path;
-    protected ?ContainerInterface $container;
-
-    private string $namespace;
-
-    public function boot(): void
-    {
-    }
-
-    public function shutdown(): void
-    {
-    }
-
-    /**
-     * This method can be overridden to register compilation passes,
-     * other extensions, ...
-     */
-    public function build(ContainerBuilder $container): void
-    {
-    }
-
     /**
      * Returns the bundle's container extension.
      *
@@ -79,15 +55,6 @@ abstract class Bundle implements BundleInterface
         return $this->extension ?: null;
     }
 
-    public function getNamespace(): string
-    {
-        if (!isset($this->namespace)) {
-            $this->parseClassName();
-        }
-
-        return $this->namespace;
-    }
-
     public function getPath(): string
     {
         if (!isset($this->path)) {
@@ -96,18 +63,6 @@ abstract class Bundle implements BundleInterface
         }
 
         return $this->path;
-    }
-
-    /**
-     * Returns the bundle name (the class short name).
-     */
-    final public function getName(): string
-    {
-        if (!isset($this->name)) {
-            $this->parseClassName();
-        }
-
-        return $this->name;
     }
 
     /**
@@ -134,17 +89,5 @@ abstract class Bundle implements BundleInterface
     protected function createContainerExtension(): ?ExtensionInterface
     {
         return class_exists($class = $this->getContainerExtensionClass()) ? new $class() : null;
-    }
-
-    private function parseClassName(): void
-    {
-        $pos = strrpos(static::class, '\\');
-        $this->namespace = false === $pos ? '' : substr(static::class, 0, $pos);
-        $this->name ??= false === $pos ? static::class : substr(static::class, $pos + 1);
-    }
-
-    public function setContainer(?ContainerInterface $container): void
-    {
-        $this->container = $container;
     }
 }
