+++ src/Symfony/Component/HttpKernel/CHANGELOG.md
@@ -26,6 +26,9 @@ CHANGELOG
  * Add `#[Serialize]` to serialize values returned by controllers
  * Add argument `$mapWhenEmpty` to `MapQueryString` and `MapRequestPayload` for always attempting denormalization with empty query and request payload
  * Deprecate `Bundle::registerCommands()`, use the `#[AsCommand]` attribute or the `console.command` service tag instead of overriding this method
+ * Deprecate `BundleInterface`, use the one from the DependencyInjection component instead
+ * Deprecate `MergeExtensionConfigurationPass`, use the one from the DependencyInjection component instead
+ * Deprecate `FileLocator`, use the one from the DependencyInjection component instead
 
 8.0
 ---
